+++ examples/lbm/cudalineupdatefunctorprototype.h
@@ -25,7 +25,7 @@ public:
     {}
 
     template<typename ACCESSOR1, typename ACCESSOR2>
-    void operator()(ACCESSOR1 accessor1, int *index1, ACCESSOR2 accessor2, int *index2) const
+    void operator()(ACCESSOR1 accessor1, long *index1, ACCESSOR2 accessor2, long *index2) const
     {
         update<CELL, ACCESSOR1, ACCESSOR2><<<dim_grid, dim_block>>>(accessor1, accessor2);
     }
@@ -45,22 +45,22 @@ public:
     {}
 
     template<typename ACCESSOR1, typename ACCESSOR2>
-    void operator()(ACCESSOR1 accessor1, int *index1, ACCESSOR2 accessor2, int *index2) const;
+    void operator()(ACCESSOR1 accessor1, long *index1, ACCESSOR2 accessor2, long *index2) const;
 
 private:
     dim3 dim_block;
     dim3 dim_grid;
 };
 
-#define IMPLEMENTATION(CELL, X1, Y1, Z1, X2, Y2, Z2)                                  \
-    template<>                                                                        \
-    template<>                                                                        \
-    void CudaLineUpdateFunctorPrototype<CellLBM>::operator()(                         \
-        LibFlatArray::soa_accessor<CELL, X1, Y1, Z1, 0> accessor1, int *index1,       \
-        LibFlatArray::soa_accessor<CELL, X2, Y2, Z2, 0> accessor2, int *index2) const \
-    {                                                                                 \
-        CudaLineUpdateFunctorPrototypeImplementation<CELL> i(dim_block, dim_grid);    \
-        i(accessor1, index1, accessor2, index2);                                      \
+#define IMPLEMENTATION(CELL, X1, Y1, Z1, X2, Y2, Z2)                    \
+    template<>                                                          \
+    template<>                                                          \
+    void CudaLineUpdateFunctorPrototype<CellLBM>::operator()(           \
+        LibFlatArray::soa_accessor<CELL, X1, Y1, Z1, 0> accessor1, long *index1, \
+        LibFlatArray::soa_accessor<CELL, X2, Y2, Z2, 0> accessor2, long *index2) const \
+    {                                                                   \
+        CudaLineUpdateFunctorPrototypeImplementation<CELL> i(dim_block, dim_grid); \
+        i(accessor1, index1, accessor2, index2);                        \
     }
